back/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
Pillow>=10.2.0
langchain-community>=0.0.38
langgraph>=0.0.62
langchain-openai>=0.0.2
langchain-core>=0.1.3
rich>=13.0.0